+++ crucible-mir/CHANGELOG.md
@@ -1,7 +1,7 @@
 # next
 
 This release supports [version
-8](https://github.com/GaloisInc/mir-json/blob/master/SCHEMA_CHANGELOG.md#8) of
+9](https://github.com/GaloisInc/mir-json/blob/master/SCHEMA_CHANGELOG.md#9) of
 `mir-json`'s schema.
 
 * Fix a translation error when translating constant slice raw pointers.
@@ -10,6 +10,7 @@ This release supports [version
   as well as a limited number of unsized types (currently, only slices).
 * Extend the override for the `atomic_xchg` intrinsic to support storing
   pointer values in addition to integer values.
+* Support translating constant trait object values.
 
 # 0.6 -- 2026-01-29