+++ crux-mir/CHANGELOG.md
@@ -1,7 +1,7 @@
 # next
 
 This release supports [version
-8](https://github.com/GaloisInc/mir-json/blob/master/SCHEMA_CHANGELOG.md#8) of
+9](https://github.com/GaloisInc/mir-json/blob/master/SCHEMA_CHANGELOG.md#9) of
 `mir-json`'s schema.
 
 # 0.12 -- 2026-01-29